+++ tuna/session_mixin.py
@@ -40,3 +40,25 @@ class SessionMixin():
   docker = Column(String(length=64),
                   nullable=False,
                   server_default="miopentuna")
+
+  def add_new_session(self, args, worker):
+    """Add new session entry"""
+    self.reason = args.label
+    self.docker = args.docker_name
+    if hasattr(args, 'arch') and args.arch:
+      self.arch = args.arch
+    else:
+      self.arch = worker.machine.arch
+
+    if hasattr(args, 'num_cu') and args.num_cu:
+      self.num_cu = args.num_cu
+    else:
+      self.num_cu = worker.machine.num_cu
+
+    if hasattr(args, 'rocm_v') and args.rocm_v:
+      self.rocm_v = args.rocm_v
+    else:
+      self.rocm_v = worker.get_rocm_v()
+
+    if hasattr(args, 'ticket') and args.ticket:
+      self.ticket = args.ticket
