+++ vars/utils.groovy
@@ -539,7 +539,7 @@ def runLint() {
           checkout scm
           def tuna_docker = docker.build("ci-tuna:${branch_id}", "--build-arg FIN_TOKEN=${FIN_TOKEN} .")
           tuna_docker.inside("") {
-            sh "cd tuna && pylint -f parseable --max-args=8 --ignore-imports=no --indent-string='  ' *.py miopen/*.py"
+            sh "cd tuna && pylint -f parseable --max-args=8 --ignore-imports=no --duplicate-code  --indent-string='  ' *.py miopen/*.py example/*.py"
           }
     }
 }
